:sparkles: [Feat] 메인 페이지 카드 UI 및 앱 구조 리팩토링
- 메인 페이지를 카드 형태로 변경하여 각 앱 접근성 향상
- weather 앱을 독립적인 하위 앱으로 분리 (/weather)
- 재사용 가능한 AppCard 컴포넌트 생성
- 반응형 그리드 레이아웃 구현
- 향후 앱 추가를 위한 확장 가능한 구조 제공
